% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-glm-reg.R
\docType{class}
\name{BinomialRegSpec-class}
\alias{BinomialRegSpec-class}
\title{Binomial GLM regression component (logit link)}
\description{
Binomial GLM regression component (logit link)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"binomial-reg"}.}

\item{\code{paramNames}}{\code{c("beta")}.}
}}

