% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-glm-reg.R
\name{BinomialRegSpec}
\alias{BinomialRegSpec}
\title{Construct a Binomial regression component spec}
\usage{
BinomialRegSpec()
}
\value{
A \code{\linkS4class{BinomialRegSpec}}.
}
\description{
Construct a Binomial regression component spec
}
\examples{
spec <- BinomialRegSpec()
}
