% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-poisson-binomial.R
\docType{class}
\name{BinomialSpec-class}
\alias{BinomialSpec-class}
\title{Binomial count component specification}
\description{
Binomial count component specification
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"binomial"}.}

\item{\code{paramNames}}{\code{c("prob")}.}
}}

