% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-poisson-binomial.R
\name{BinomialSpec}
\alias{BinomialSpec}
\title{Construct a Binomial component spec}
\usage{
BinomialSpec()
}
\value{
A \code{\linkS4class{BinomialSpec}}.
}
\description{
Construct a Binomial component spec
}
\examples{
spec <- BinomialSpec()
}
