% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-glm-reg.R,
%   R/dist-gmsnburr.R, R/dist-normal-reg.R, R/dist-heavytail-reg.R,
%   R/dist-msnburr.R, R/dist-mv-reg.R, R/dist-normal-mv.R,
%   R/dist-normal-gamma-mv.R, R/dist-normal-uv.R, R/dist-normal-gamma.R,
%   R/dist-poisson-binomial.R, R/dist-student-t-mv.R, R/dist-student-t.R
\name{buildConstants}
\alias{buildConstants}
\alias{buildConstants,PoissonRegSpec-method}
\alias{buildConstants,BinomialRegSpec-method}
\alias{buildConstants,GMSNBurrUvSpec-method}
\alias{buildConstants,NormalRegSpec-method}
\alias{buildConstants,StudentTRegSpec-method}
\alias{buildConstants,NormalGammaRegSpec-method}
\alias{buildConstants,MSNBurrUvSpec-method}
\alias{buildConstants,MSNBurr2aUvSpec-method}
\alias{buildConstants,NormalMvRegSpec-method}
\alias{buildConstants,StudentTMvRegSpec-method}
\alias{buildConstants,NormalGammaMvRegSpec-method}
\alias{buildConstants,NormalMvSpec-method}
\alias{buildConstants,NormalGammaMvSpec-method}
\alias{buildConstants,NormalUvSpec-method}
\alias{buildConstants,NormalGammaUvSpec-method}
\alias{buildConstants,PoissonSpec-method}
\alias{buildConstants,BinomialSpec-method}
\alias{buildConstants,StudentTMvSpec-method}
\alias{buildConstants,StudentTUvSpec-method}
\title{Assemble distribution-specific NIMBLE constants for the DPM engine}
\usage{
buildConstants(spec, prior, n, ...)

\S4method{buildConstants}{PoissonRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{BinomialRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{GMSNBurrUvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{StudentTRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalGammaRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{MSNBurrUvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{MSNBurr2aUvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalMvRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{StudentTMvRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalGammaMvRegSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalMvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalGammaMvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalUvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{NormalGammaUvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{PoissonSpec}(spec, prior, n, ...)

\S4method{buildConstants}{BinomialSpec}(spec, prior, n, ...)

\S4method{buildConstants}{StudentTMvSpec}(spec, prior, n, ...)

\S4method{buildConstants}{StudentTUvSpec}(spec, prior, n, ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{prior}{The prior list (from \code{\link{defaultPrior}}).}

\item{n}{Integer number of observations.}

\item{...}{Reserved for methods.}
}
\value{
A named list of constants (excluding the concentration hyperprior,
  which the engine appends).
}
\description{
Assemble distribution-specific NIMBLE constants for the DPM engine
}
\section{Functions}{
\itemize{
\item \code{buildConstants(PoissonRegSpec)}: Poisson regression constants.

\item \code{buildConstants(BinomialRegSpec)}: Binomial regression constants plus \code{size}.

\item \code{buildConstants(GMSNBurrUvSpec)}: GMSNBurr constants.

\item \code{buildConstants(NormalRegSpec)}: Regression constants (design matrix \code{X},
number of predictors \code{p}, and the NIG hyperparameters).

\item \code{buildConstants(StudentTRegSpec)}: NIG g-prior constants plus \code{df}.

\item \code{buildConstants(NormalGammaRegSpec)}: NIG g-prior constants plus \code{df}.

\item \code{buildConstants(MSNBurrUvSpec)}: MSNBurr constants.

\item \code{buildConstants(MSNBurr2aUvSpec)}: MSNBurr-IIa constants.

\item \code{buildConstants(NormalMvRegSpec)}: Multivariate-regression constants.

\item \code{buildConstants(StudentTMvRegSpec)}: Multivariate-regression constants plus \code{df}.

\item \code{buildConstants(NormalGammaMvRegSpec)}: Multivariate-regression constants plus \code{df}.

\item \code{buildConstants(NormalMvSpec)}: Multivariate Normal-Inverse-Wishart constants
(includes the dimension \code{d}, mean vector \code{mu0} and scale matrix
\code{S0}).

\item \code{buildConstants(NormalGammaMvSpec)}: Normal-Inverse-Wishart constants plus \code{df}.

\item \code{buildConstants(NormalUvSpec)}: Univariate Normal-Inverse-Gamma constants.

\item \code{buildConstants(NormalGammaUvSpec)}: Normal-Inverse-Gamma constants plus \code{df}.

\item \code{buildConstants(PoissonSpec)}: Poisson Gamma-prior constants.

\item \code{buildConstants(BinomialSpec)}: Binomial Beta-prior constants plus \code{size}.

\item \code{buildConstants(StudentTMvSpec)}: Normal-Inverse-Wishart constants plus \code{df}.

\item \code{buildConstants(StudentTUvSpec)}: Student-t constants (location/precision prior + df).

}}
\keyword{internal}
