% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-fssn.R,
%   R/dist-fossep.R, R/dist-fsst.R, R/dist-glm-reg.R, R/dist-gmsnburr.R,
%   R/dist-normal-reg.R, R/dist-jfst.R, R/dist-sep.R, R/dist-lep.R,
%   R/dist-msnburr.R, R/dist-mv-reg.R, R/dist-normal-mv.R, R/dist-normal-uv.R,
%   R/dist-poisson-binomial.R, R/dist-skewnormal-mv.R,
%   R/dist-skewistudent-mv.R, R/dist-student-t.R
\name{buildDataList}
\alias{buildDataList}
\alias{buildDataList,FSSNUvSpec-method}
\alias{buildDataList,FOSSEPUvSpec-method}
\alias{buildDataList,FSSTUvSpec-method}
\alias{buildDataList,PoissonRegSpec-method}
\alias{buildDataList,BinomialRegSpec-method}
\alias{buildDataList,GMSNBurrUvSpec-method}
\alias{buildDataList,NormalRegSpec-method}
\alias{buildDataList,JFSTUvSpec-method}
\alias{buildDataList,SEPUvSpec-method}
\alias{buildDataList,LEPUvSpec-method}
\alias{buildDataList,MSNBurrUvSpec-method}
\alias{buildDataList,MSNBurr2aUvSpec-method}
\alias{buildDataList,NormalMvRegSpec-method}
\alias{buildDataList,NormalMvSpec-method}
\alias{buildDataList,NormalUvSpec-method}
\alias{buildDataList,PoissonSpec-method}
\alias{buildDataList,BinomialSpec-method}
\alias{buildDataList,SkewNormalMvSpec-method}
\alias{buildDataList,SkewIStudentMvSpec-method}
\alias{buildDataList,StudentTUvSpec-method}
\title{Shape the observed data into the NIMBLE \code{data} list}
\usage{
buildDataList(spec, data, ...)

\S4method{buildDataList}{FSSNUvSpec}(spec, data, ...)

\S4method{buildDataList}{FOSSEPUvSpec}(spec, data, ...)

\S4method{buildDataList}{FSSTUvSpec}(spec, data, ...)

\S4method{buildDataList}{PoissonRegSpec}(spec, data, ...)

\S4method{buildDataList}{BinomialRegSpec}(spec, data, ...)

\S4method{buildDataList}{GMSNBurrUvSpec}(spec, data, ...)

\S4method{buildDataList}{NormalRegSpec}(spec, data, ...)

\S4method{buildDataList}{JFSTUvSpec}(spec, data, ...)

\S4method{buildDataList}{SEPUvSpec}(spec, data, ...)

\S4method{buildDataList}{LEPUvSpec}(spec, data, ...)

\S4method{buildDataList}{MSNBurrUvSpec}(spec, data, ...)

\S4method{buildDataList}{MSNBurr2aUvSpec}(spec, data, ...)

\S4method{buildDataList}{NormalMvRegSpec}(spec, data, ...)

\S4method{buildDataList}{NormalMvSpec}(spec, data, ...)

\S4method{buildDataList}{NormalUvSpec}(spec, data, ...)

\S4method{buildDataList}{PoissonSpec}(spec, data, ...)

\S4method{buildDataList}{BinomialSpec}(spec, data, ...)

\S4method{buildDataList}{SkewNormalMvSpec}(spec, data, ...)

\S4method{buildDataList}{SkewIStudentMvSpec}(spec, data, ...)

\S4method{buildDataList}{StudentTUvSpec}(spec, data, ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{data}{The observed data (vector or matrix).}

\item{...}{Reserved for methods.}
}
\value{
A named list, typically \code{list(y = ...)}.
}
\description{
Shape the observed data into the NIMBLE \code{data} list
}
\section{Functions}{
\itemize{
\item \code{buildDataList(FSSNUvSpec)}: Univariate response.

\item \code{buildDataList(FOSSEPUvSpec)}: Univariate response.

\item \code{buildDataList(FSSTUvSpec)}: Univariate response.

\item \code{buildDataList(PoissonRegSpec)}: Response and design matrix.

\item \code{buildDataList(BinomialRegSpec)}: Response and design matrix.

\item \code{buildDataList(GMSNBurrUvSpec)}: Univariate response.

\item \code{buildDataList(NormalRegSpec)}: Response vector for the regression mixture.

\item \code{buildDataList(JFSTUvSpec)}: Univariate response.

\item \code{buildDataList(SEPUvSpec)}: Univariate response.

\item \code{buildDataList(LEPUvSpec)}: Univariate response.

\item \code{buildDataList(MSNBurrUvSpec)}: Univariate response.

\item \code{buildDataList(MSNBurr2aUvSpec)}: Univariate response.

\item \code{buildDataList(NormalMvRegSpec)}: Matrix response.

\item \code{buildDataList(NormalMvSpec)}: Multivariate data matrix (one row per observation).

\item \code{buildDataList(NormalUvSpec)}: Univariate data vector.

\item \code{buildDataList(PoissonSpec)}: Count data vector.

\item \code{buildDataList(BinomialSpec)}: Count data vector.

\item \code{buildDataList(SkewNormalMvSpec)}: Multivariate response matrix.

\item \code{buildDataList(SkewIStudentMvSpec)}: Multivariate response matrix.

\item \code{buildDataList(StudentTUvSpec)}: Univariate data vector.

}}
\keyword{internal}
