% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-fssn.R,
%   R/dist-fossep.R, R/dist-fsst.R, R/dist-glm-reg.R, R/dist-gmsnburr.R,
%   R/dist-normal-reg.R, R/dist-heavytail-reg.R, R/dist-jfst.R, R/dist-sep.R,
%   R/dist-lep.R, R/dist-msnburr.R, R/dist-mv-reg.R, R/dist-normal-mv.R,
%   R/dist-normal-gamma-mv.R, R/dist-normal-uv.R, R/dist-normal-gamma.R,
%   R/dist-poisson-binomial.R, R/dist-skewnormal-mv.R,
%   R/dist-skewnormal-mv-o.R, R/dist-skewistudent-mv.R,
%   R/dist-skewistudent-mv-o.R, R/dist-skewnormal-mv-og.R,
%   R/dist-skewistudent-mv-og.R, R/dist-student-t-mv.R, R/dist-student-t.R,
%   R/engine-fixedk.R, R/engine-hmm.R, R/engine-mrf.R
\name{buildModelCode}
\alias{buildModelCode}
\alias{buildModelCode,FSSNUvSpec,FixedKEngine-method}
\alias{buildModelCode,FSSNUvSpec,DPMEngine-method}
\alias{buildModelCode,FOSSEPUvSpec,FixedKEngine-method}
\alias{buildModelCode,FOSSEPUvSpec,DPMEngine-method}
\alias{buildModelCode,FSSTUvSpec,FixedKEngine-method}
\alias{buildModelCode,FSSTUvSpec,DPMEngine-method}
\alias{buildModelCode,PoissonRegSpec,DPMEngine-method}
\alias{buildModelCode,PoissonRegSpec,FixedKEngine-method}
\alias{buildModelCode,BinomialRegSpec,DPMEngine-method}
\alias{buildModelCode,BinomialRegSpec,FixedKEngine-method}
\alias{buildModelCode,GMSNBurrUvSpec,FixedKEngine-method}
\alias{buildModelCode,GMSNBurrUvSpec,DPMEngine-method}
\alias{buildModelCode,NormalRegSpec,DPMEngine-method}
\alias{buildModelCode,StudentTRegSpec,DPMEngine-method}
\alias{buildModelCode,StudentTRegSpec,FixedKEngine-method}
\alias{buildModelCode,NormalGammaRegSpec,DPMEngine-method}
\alias{buildModelCode,NormalGammaRegSpec,FixedKEngine-method}
\alias{buildModelCode,JFSTUvSpec,FixedKEngine-method}
\alias{buildModelCode,JFSTUvSpec,DPMEngine-method}
\alias{buildModelCode,SEPUvSpec,FixedKEngine-method}
\alias{buildModelCode,SEPUvSpec,DPMEngine-method}
\alias{buildModelCode,LEPUvSpec,FixedKEngine-method}
\alias{buildModelCode,LEPUvSpec,DPMEngine-method}
\alias{buildModelCode,MSNBurrUvSpec,FixedKEngine-method}
\alias{buildModelCode,MSNBurr2aUvSpec,FixedKEngine-method}
\alias{buildModelCode,MSNBurrUvSpec,DPMEngine-method}
\alias{buildModelCode,MSNBurr2aUvSpec,DPMEngine-method}
\alias{buildModelCode,NormalMvRegSpec,DPMEngine-method}
\alias{buildModelCode,NormalMvRegSpec,FixedKEngine-method}
\alias{buildModelCode,StudentTMvRegSpec,DPMEngine-method}
\alias{buildModelCode,StudentTMvRegSpec,FixedKEngine-method}
\alias{buildModelCode,NormalGammaMvRegSpec,DPMEngine-method}
\alias{buildModelCode,NormalGammaMvRegSpec,FixedKEngine-method}
\alias{buildModelCode,NormalMvSpec,DPMEngine-method}
\alias{buildModelCode,NormalGammaMvSpec,DPMEngine-method}
\alias{buildModelCode,NormalUvSpec,DPMEngine-method}
\alias{buildModelCode,NormalGammaUvSpec,DPMEngine-method}
\alias{buildModelCode,PoissonSpec,DPMEngine-method}
\alias{buildModelCode,BinomialSpec,DPMEngine-method}
\alias{buildModelCode,SkewNormalMvSpec,FixedKEngine-method}
\alias{buildModelCode,SkewNormalMvSpec,DPMEngine-method}
\alias{buildModelCode,SkewNormalMvOSpec,FixedKEngine-method}
\alias{buildModelCode,SkewNormalMvOSpec,DPMEngine-method}
\alias{buildModelCode,SkewIStudentMvSpec,FixedKEngine-method}
\alias{buildModelCode,SkewIStudentMvSpec,DPMEngine-method}
\alias{buildModelCode,SkewIStudentMvOSpec,FixedKEngine-method}
\alias{buildModelCode,SkewIStudentMvOSpec,DPMEngine-method}
\alias{buildModelCode,SkewNormalMvOGenSpec,FixedKEngine-method}
\alias{buildModelCode,SkewNormalMvOGenSpec,DPMEngine-method}
\alias{buildModelCode,SkewIStudentMvOGenSpec,FixedKEngine-method}
\alias{buildModelCode,SkewIStudentMvOGenSpec,DPMEngine-method}
\alias{buildModelCode,StudentTMvSpec,DPMEngine-method}
\alias{buildModelCode,StudentTUvSpec,DPMEngine-method}
\alias{buildModelCode,NormalUvSpec,FixedKEngine-method}
\alias{buildModelCode,NormalMvSpec,FixedKEngine-method}
\alias{buildModelCode,NormalRegSpec,FixedKEngine-method}
\alias{buildModelCode,StudentTUvSpec,FixedKEngine-method}
\alias{buildModelCode,NormalGammaUvSpec,FixedKEngine-method}
\alias{buildModelCode,NormalGammaMvSpec,FixedKEngine-method}
\alias{buildModelCode,StudentTMvSpec,FixedKEngine-method}
\alias{buildModelCode,PoissonSpec,FixedKEngine-method}
\alias{buildModelCode,BinomialSpec,FixedKEngine-method}
\alias{buildModelCode,NormalUvSpec,HMMEngine-method}
\alias{buildModelCode,StudentTUvSpec,HMMEngine-method}
\alias{buildModelCode,PoissonSpec,HMMEngine-method}
\alias{buildModelCode,MSNBurrUvSpec,HMMEngine-method}
\alias{buildModelCode,MSNBurr2aUvSpec,HMMEngine-method}
\alias{buildModelCode,GMSNBurrUvSpec,HMMEngine-method}
\alias{buildModelCode,DistributionSpec,MRFEngine-method}
\title{Build the NIMBLE model code for a (distribution, engine) pair}
\usage{
buildModelCode(spec, engine, n, L, ...)

\S4method{buildModelCode}{FSSNUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{FSSNUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{FOSSEPUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{FOSSEPUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{FSSTUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{FSSTUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{BinomialRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{BinomialRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{GMSNBurrUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{GMSNBurrUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{JFSTUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{JFSTUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SEPUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SEPUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{LEPUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{LEPUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurrUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurr2aUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurrUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurr2aUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalMvRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalMvRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTMvRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTMvRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaMvRegSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaMvRegSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalMvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaMvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{BinomialSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvOSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvOSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvOSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvOSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvOGenSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewNormalMvOGenSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvOGenSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{SkewIStudentMvOGenSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTMvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTUvSpec,DPMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalMvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalRegSpec,FixedKEngine}(spec, engine, n, L, re = FALSE, ...)

\S4method{buildModelCode}{StudentTUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaUvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalGammaMvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTMvSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{BinomialSpec,FixedKEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{NormalUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{StudentTUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurrUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurr2aUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{GMSNBurrUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{MSNBurrUvSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{PoissonSpec,HMMEngine}(spec, engine, n, L, ...)

\S4method{buildModelCode}{DistributionSpec,MRFEngine}(spec, engine, n, L, ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{engine}{An \code{\linkS4class{EngineConfig}}.}

\item{n}{Integer, number of observations.}

\item{L}{Integer, cluster-parameter truncation length (\code{= K_max}).}

\item{...}{Reserved for methods.}

\item{re}{Logical; regression + FixedK method only -- when \code{TRUE},
emit the random-intercept variant of the model code (sum-to-zero
\code{b} offsets plus \code{tauRE}).}
}
\value{
A list with elements \code{code} (a \code{nimbleCode} object),
  \code{monitors} (character vector), and \code{paramNodes}
  (named character vector mapping logical parameter names to model nodes).
}
\description{
Dispatches on BOTH the component distribution and the engine. This is the
extensibility seam described in adding a new
distribution to the DPM engine means adding one method here, not editing the
engine or the spec base class.
}
\section{Functions}{
\itemize{
\item \code{buildModelCode(spec = FSSNUvSpec, engine = FixedKEngine)}: FSSN finite mixture (fixed K).

\item \code{buildModelCode(spec = FSSNUvSpec, engine = DPMEngine)}: FSSN DPM mixture.

\item \code{buildModelCode(spec = FOSSEPUvSpec, engine = FixedKEngine)}: FOSSEP finite mixture (fixed K).

\item \code{buildModelCode(spec = FOSSEPUvSpec, engine = DPMEngine)}: FOSSEP DPM mixture.

\item \code{buildModelCode(spec = FSSTUvSpec, engine = FixedKEngine)}: FSST finite mixture (fixed K).

\item \code{buildModelCode(spec = FSSTUvSpec, engine = DPMEngine)}: FSST DPM mixture.

\item \code{buildModelCode(spec = PoissonRegSpec, engine = DPMEngine)}: Poisson GLM regression DPM code (log link).

\item \code{buildModelCode(spec = PoissonRegSpec, engine = FixedKEngine)}: Poisson GLM regression fixed-K code (log link).

\item \code{buildModelCode(spec = BinomialRegSpec, engine = DPMEngine)}: Binomial GLM regression DPM code (logit link).

\item \code{buildModelCode(spec = BinomialRegSpec, engine = FixedKEngine)}: Binomial GLM regression fixed-K code (logit link).

\item \code{buildModelCode(spec = GMSNBurrUvSpec, engine = FixedKEngine)}: GMSNBurr finite mixture (fixed K).

\item \code{buildModelCode(spec = GMSNBurrUvSpec, engine = DPMEngine)}: GMSNBurr DPM mixture.

\item \code{buildModelCode(spec = NormalRegSpec, engine = DPMEngine)}: DPM mixture-of-linear-regressions model code
  (dCRP) with a conjugate Normal-Inverse-Gamma cluster prior and constant
  (CRP) gating.

Builds NIMBLE code for
\deqn{y_i \sim N(x_i^\top \beta_{\xi_i}, s^2_{\xi_i}), \quad
      \xi_{1:n} \sim CRP(\alpha, n),}
with \eqn{s^2_j \sim InvGamma(nu_0, s_0)} and
\eqn{\beta_j \mid s^2_j \sim N_p(b_0, s^2_j B_0)}. The number of predictors
\eqn{p} and the design matrix \eqn{X} are constants.

\item \code{buildModelCode(spec = StudentTRegSpec, engine = DPMEngine)}: Student-t regression DPM code (direct t density;
the scale enters as precision \code{tau = 1 / s2}).

\item \code{buildModelCode(spec = StudentTRegSpec, engine = FixedKEngine)}: Student-t regression fixed-K code.

\item \code{buildModelCode(spec = NormalGammaRegSpec, engine = DPMEngine)}: Normal-Gamma regression DPM code (scale mixture;
conjugate coefficients conditional on \code{omega}).

\item \code{buildModelCode(spec = NormalGammaRegSpec, engine = FixedKEngine)}: Normal-Gamma regression fixed-K code.

\item \code{buildModelCode(spec = JFSTUvSpec, engine = FixedKEngine)}: JFST finite mixture (fixed K).

\item \code{buildModelCode(spec = JFSTUvSpec, engine = DPMEngine)}: JFST DPM mixture.

\item \code{buildModelCode(spec = SEPUvSpec, engine = FixedKEngine)}: SEP finite mixture (fixed K).

\item \code{buildModelCode(spec = SEPUvSpec, engine = DPMEngine)}: SEP DPM mixture.

\item \code{buildModelCode(spec = LEPUvSpec, engine = FixedKEngine)}: LEP finite mixture (fixed K).

\item \code{buildModelCode(spec = LEPUvSpec, engine = DPMEngine)}: LEP DPM mixture.

\item \code{buildModelCode(spec = MSNBurrUvSpec, engine = FixedKEngine)}: MSNBurr finite mixture (fixed K).

\item \code{buildModelCode(spec = MSNBurr2aUvSpec, engine = FixedKEngine)}: MSNBurr-IIa finite mixture (fixed K).

\item \code{buildModelCode(spec = MSNBurrUvSpec, engine = DPMEngine)}: MSNBurr DPM mixture.

\item \code{buildModelCode(spec = MSNBurr2aUvSpec, engine = DPMEngine)}: MSNBurr-IIa DPM mixture.

\item \code{buildModelCode(spec = NormalMvRegSpec, engine = DPMEngine)}: Multivariate-response Normal regression DPM code.

\item \code{buildModelCode(spec = NormalMvRegSpec, engine = FixedKEngine)}: Multivariate-response Normal regression fixed-K.

\item \code{buildModelCode(spec = StudentTMvRegSpec, engine = DPMEngine)}: Multivariate-response Student-t regression DPM code
(direct multivariate-t kernel).

\item \code{buildModelCode(spec = StudentTMvRegSpec, engine = FixedKEngine)}: Multivariate-response Student-t regression fixed-K.

\item \code{buildModelCode(spec = NormalGammaMvRegSpec, engine = DPMEngine)}: Multivariate-response Normal-Gamma regression DPM
code (scale mixture; conjugate cluster updates).

\item \code{buildModelCode(spec = NormalGammaMvRegSpec, engine = FixedKEngine)}: Multivariate-response Normal-Gamma regression
fixed-K code.

\item \code{buildModelCode(spec = NormalMvSpec, engine = DPMEngine)}: Multivariate Gaussian DPM model code (dCRP) with a
  conjugate Normal-Inverse-Wishart cluster base measure.

Builds NIMBLE code for
\deqn{y_i \sim N_d(\mu_{\xi_i}, \Sigma_{\xi_i}), \quad
      \xi_{1:n} \sim CRP(\alpha, n),}
with \eqn{\Sigma_j \sim InvWishart(S_0, df_0)} and
\eqn{\mu_j \sim N_d(\mu_0, \Sigma_j / \kappa_0)}. The dimension \eqn{d} is a
constant so the index ranges \code{1:d} expand at model-build time.

\item \code{buildModelCode(spec = NormalGammaMvSpec, engine = DPMEngine)}: Multivariate Normal-Gamma scale-mixture DPM code
with a per-observation latent precision multiplier \code{omega}.

\item \code{buildModelCode(spec = NormalUvSpec, engine = DPMEngine)}: Univariate Gaussian DPM model code (dCRP).

Builds the NIMBLE code for
\deqn{y_i \sim N(\mu_{\xi_i}, s^2_{\xi_i}), \quad
      \xi_{1:n} \sim CRP(\alpha, n),}
with a conjugate Normal-Inverse-Gamma cluster prior and a Gamma hyperprior
on \eqn{\alpha}. Cluster-parameter vectors have length \code{L = K_max}
(NIMBLE's exact truncation; the sampler stays proper as long as the number
of occupied clusters is strictly below \code{L}).

\item \code{buildModelCode(spec = NormalGammaUvSpec, engine = DPMEngine)}: Normal-Gamma scale-mixture DPM model code (dCRP)
with a per-observation latent precision multiplier \code{omega}.

\item \code{buildModelCode(spec = PoissonSpec, engine = DPMEngine)}: Poisson DPM model code (Gamma-Poisson conjugate).

\item \code{buildModelCode(spec = BinomialSpec, engine = DPMEngine)}: Binomial DPM model code (Beta-Binomial conjugate).

\item \code{buildModelCode(spec = SkewNormalMvSpec, engine = FixedKEngine)}: Skew-mv-Normal finite mixture (fixed K).

\item \code{buildModelCode(spec = SkewNormalMvSpec, engine = DPMEngine)}: Skew-mv-Normal DPM mixture.

\item \code{buildModelCode(spec = SkewNormalMvOSpec, engine = FixedKEngine)}: Skew-mv-Normal-O finite mixture (fixed K).

\item \code{buildModelCode(spec = SkewNormalMvOSpec, engine = DPMEngine)}: Skew-mv-Normal-O DPM mixture.

\item \code{buildModelCode(spec = SkewIStudentMvSpec, engine = FixedKEngine)}: Skew-mv-IStudent finite mixture (fixed K).

\item \code{buildModelCode(spec = SkewIStudentMvSpec, engine = DPMEngine)}: Skew-mv-IStudent DPM mixture.

\item \code{buildModelCode(spec = SkewIStudentMvOSpec, engine = FixedKEngine)}: Skew-mv-IStudent-O finite mixture (fixed K).

\item \code{buildModelCode(spec = SkewIStudentMvOSpec, engine = DPMEngine)}: Skew-mv-IStudent-O DPM mixture.

\item \code{buildModelCode(spec = SkewNormalMvOGenSpec, engine = FixedKEngine)}: General-m skew-mv-Normal-O finite mixture.

\item \code{buildModelCode(spec = SkewNormalMvOGenSpec, engine = DPMEngine)}: General-m skew-mv-Normal-O DPM mixture.

\item \code{buildModelCode(spec = SkewIStudentMvOGenSpec, engine = FixedKEngine)}: General-m skew-mv-IStudent-O finite mixture.

\item \code{buildModelCode(spec = SkewIStudentMvOGenSpec, engine = DPMEngine)}: General-m skew-mv-IStudent-O DPM mixture.

\item \code{buildModelCode(spec = StudentTMvSpec, engine = DPMEngine)}: Multivariate Student-t DPM code using the
user-defined \code{dmvt_nimix} kernel with a Normal-Inverse-Wishart cluster
prior.

\item \code{buildModelCode(spec = StudentTUvSpec, engine = DPMEngine)}: Univariate Student-t DPM model code (dCRP). The
t density is evaluated directly; df is a constant.

\item \code{buildModelCode(spec = NormalUvSpec, engine = FixedKEngine)}: Univariate Gaussian finite-mixture code (fixed K).

\item \code{buildModelCode(spec = NormalMvSpec, engine = FixedKEngine)}: Multivariate Gaussian finite-mixture code (fixed K).

\item \code{buildModelCode(spec = NormalRegSpec, engine = FixedKEngine)}: Mixture-of-linear-regressions finite-mixture code
(fixed K).

\item \code{buildModelCode(spec = StudentTUvSpec, engine = FixedKEngine)}: Univariate Student-t finite-mixture code (fixed K).

\item \code{buildModelCode(spec = NormalGammaUvSpec, engine = FixedKEngine)}: Univariate Normal-Gamma (scale-mixture Student-t)
finite-mixture code (fixed K).

\item \code{buildModelCode(spec = NormalGammaMvSpec, engine = FixedKEngine)}: Multivariate Normal-Gamma scale-mixture
finite-mixture code (fixed K).

\item \code{buildModelCode(spec = StudentTMvSpec, engine = FixedKEngine)}: Multivariate Student-t finite-mixture code
(fixed K) using the user-defined \code{dmvt_nimix} kernel.

\item \code{buildModelCode(spec = PoissonSpec, engine = FixedKEngine)}: Poisson finite-mixture code (fixed K).

\item \code{buildModelCode(spec = BinomialSpec, engine = FixedKEngine)}: Binomial finite-mixture code (fixed K).

\item \code{buildModelCode(spec = NormalUvSpec, engine = HMMEngine)}: Univariate Gaussian regime-switching HMM code
(states marginalised by the forward algorithm; no allocation node).

\item \code{buildModelCode(spec = StudentTUvSpec, engine = HMMEngine)}: Univariate Student-t regime-switching HMM code
(fixed df; states marginalised by the forward algorithm).

\item \code{buildModelCode(spec = PoissonSpec, engine = HMMEngine)}: Poisson regime-switching HMM code (count data;
states marginalised by the forward algorithm).

\item \code{buildModelCode(spec = MSNBurrUvSpec, engine = HMMEngine)}: MSNBurr (neo-normal, skewed) regime-switching
HMM code (states marginalised by the forward algorithm).

\item \code{buildModelCode(spec = MSNBurr2aUvSpec, engine = HMMEngine)}: MSNBurr-IIa regime-switching HMM code.

\item \code{buildModelCode(spec = GMSNBurrUvSpec, engine = HMMEngine)}: GMSNBurr (four-parameter, skewed) regime-switching
HMM code.

\item \code{buildModelCode(spec = MSNBurrUvSpec, engine = HMMEngine)}: MSNBurr regime-switching HMM code (skewed
regimes; states marginalised by the forward algorithm).

\item \code{buildModelCode(spec = PoissonSpec, engine = HMMEngine)}: Poisson regime-switching HMM code (count
regimes; states marginalised by the forward algorithm).

\item \code{buildModelCode(spec = DistributionSpec, engine = MRFEngine)}: Default MRF kernel for any family: the family's
fixed-K kernel with the Dirichlet-categorical label layer replaced by the
joint Potts node (derived mechanically; see the spatial design notes).

}}
\references{
Hurn et al. (2003) \doi{10.1198/1061860031329};
de Valpine et al. (2017) \doi{10.1080/10618600.2016.1172487}.

Zhang et al. (2004) \doi{10.1023/B:STCO.0000039481.32735.0c};
de Valpine et al. (2017) \doi{10.1080/10618600.2016.1172487}.

Neal, R.M. (2000) \doi{10.1080/10618600.2000.10474879};
de Valpine et al. (2017) \doi{10.1080/10618600.2016.1172487}.
}
