% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MixtureModel.R
\docType{class}
\name{ClusterModel-class}
\alias{ClusterModel-class}
\title{Mixture clustering model (for nimixClust)}
\description{
Mixture clustering model (for nimixClust)
}
\section{Slots}{

\describe{
\item{\code{data}}{Numeric vector (univariate) or numeric matrix with one row per
observation and one column per dimension (multivariate, v0.2.0).}
}}

