% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-fssn.R,
%   R/dist-fossep.R, R/dist-fsst.R, R/dist-glm-reg.R, R/dist-gmsnburr.R,
%   R/dist-normal-reg.R, R/dist-heavytail-reg.R, R/dist-jfst.R, R/dist-sep.R,
%   R/dist-lep.R, R/dist-msnburr.R, R/dist-mv-reg.R, R/dist-normal-mv.R,
%   R/dist-normal-gamma-mv.R, R/dist-normal-uv.R, R/dist-normal-gamma.R,
%   R/dist-poisson-binomial.R, R/dist-skewnormal-mv.R,
%   R/dist-skewnormal-mv-o.R, R/dist-skewistudent-mv.R,
%   R/dist-skewistudent-mv-o.R, R/dist-skewnormal-mv-og.R,
%   R/dist-skewistudent-mv-og.R, R/dist-student-t-mv.R, R/dist-student-t.R
\name{defaultPrior}
\alias{defaultPrior}
\alias{defaultPrior,FSSNUvSpec-method}
\alias{defaultPrior,FOSSEPUvSpec-method}
\alias{defaultPrior,FSSTUvSpec-method}
\alias{defaultPrior,PoissonRegSpec-method}
\alias{defaultPrior,BinomialRegSpec-method}
\alias{defaultPrior,GMSNBurrUvSpec-method}
\alias{defaultPrior,NormalRegSpec-method}
\alias{defaultPrior,StudentTRegSpec-method}
\alias{defaultPrior,NormalGammaRegSpec-method}
\alias{defaultPrior,JFSTUvSpec-method}
\alias{defaultPrior,SEPUvSpec-method}
\alias{defaultPrior,LEPUvSpec-method}
\alias{defaultPrior,MSNBurrUvSpec-method}
\alias{defaultPrior,MSNBurr2aUvSpec-method}
\alias{defaultPrior,NormalMvRegSpec-method}
\alias{defaultPrior,StudentTMvRegSpec-method}
\alias{defaultPrior,NormalGammaMvRegSpec-method}
\alias{defaultPrior,NormalMvSpec-method}
\alias{defaultPrior,NormalGammaMvSpec-method}
\alias{defaultPrior,NormalUvSpec-method}
\alias{defaultPrior,NormalGammaUvSpec-method}
\alias{defaultPrior,PoissonSpec-method}
\alias{defaultPrior,BinomialSpec-method}
\alias{defaultPrior,SkewNormalMvSpec-method}
\alias{defaultPrior,SkewNormalMvOSpec-method}
\alias{defaultPrior,SkewIStudentMvSpec-method}
\alias{defaultPrior,SkewIStudentMvOSpec-method}
\alias{defaultPrior,SkewNormalMvOGenSpec-method}
\alias{defaultPrior,SkewIStudentMvOGenSpec-method}
\alias{defaultPrior,StudentTMvSpec-method}
\alias{defaultPrior,StudentTUvSpec-method}
\title{Build a data-scaled default prior for a distribution}
\usage{
defaultPrior(spec, data, control = list(), ...)

\S4method{defaultPrior}{FSSNUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{FOSSEPUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{FSSTUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{PoissonRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{BinomialRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{GMSNBurrUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{StudentTRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalGammaRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{JFSTUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SEPUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{LEPUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{MSNBurrUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{MSNBurr2aUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalMvRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{StudentTMvRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalGammaMvRegSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalMvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalGammaMvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{NormalGammaUvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{PoissonSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{BinomialSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewNormalMvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewNormalMvOSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewIStudentMvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewIStudentMvOSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewNormalMvOGenSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{SkewIStudentMvOGenSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{StudentTMvSpec}(spec, data, control = list(), ...)

\S4method{defaultPrior}{StudentTUvSpec}(spec, data, control = list(), ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{data}{Numeric data used to scale the prior.}

\item{control}{A named list of user overrides (merged over the defaults).}

\item{...}{Reserved for methods.}
}
\value{
A named list of prior hyperparameters.
}
\description{
Returns a named list of prior hyperparameters scaled to the observed data,
following the weakly-informative, data-scaled philosophy in project
knowledge (priors for location parameters must not be made
arbitrarily vague when \code{K_max} is large).
}
\section{Functions}{
\itemize{
\item \code{defaultPrior(FSSNUvSpec)}: Data-scaled FSSN prior (log-normal skewness).

\item \code{defaultPrior(FOSSEPUvSpec)}: Data-scaled FOSSEP prior.

\item \code{defaultPrior(FSSTUvSpec)}: Data-scaled FSST prior.

\item \code{defaultPrior(PoissonRegSpec)}: g-prior on the coefficients.

\item \code{defaultPrior(BinomialRegSpec)}: g-prior on the coefficients; needs \code{size}.

\item \code{defaultPrior(GMSNBurrUvSpec)}: Data-scaled GMSNBurr prior.

\item \code{defaultPrior(NormalRegSpec)}: Data-scaled Normal-Inverse-Gamma g-prior for the
  regression component. Requires the design matrix in \code{control$X}.

Control overrides: \code{g} (g-prior factor; prior covariance of \code{beta}
is \code{s2 * g * solve(crossprod(X))}, default \code{g = n}, the unit-information
prior) and \code{nu0} (InvGamma shape, default 3, must exceed 2 for a finite
prior variance).

\item \code{defaultPrior(StudentTRegSpec)}: NIG g-prior plus a fixed \code{df} (default 4, > 2).

\item \code{defaultPrior(NormalGammaRegSpec)}: NIG g-prior plus a fixed \code{df} (default 4, > 2).

\item \code{defaultPrior(JFSTUvSpec)}: Data-scaled JFST prior (symmetric in alpha/theta).

\item \code{defaultPrior(SEPUvSpec)}: Data-scaled SEP prior.

\item \code{defaultPrior(LEPUvSpec)}: Data-scaled LEP prior.

\item \code{defaultPrior(MSNBurrUvSpec)}: Data-scaled MSNBurr prior (location/scale/shape).

\item \code{defaultPrior(MSNBurr2aUvSpec)}: Data-scaled MSNBurr-IIa prior.

\item \code{defaultPrior(NormalMvRegSpec)}: Inverse-Wishart on Sigma + matrix-normal coefficients.

\item \code{defaultPrior(StudentTMvRegSpec)}: Inverse-Wishart + matrix-normal plus a fixed \code{df}.

\item \code{defaultPrior(NormalGammaMvRegSpec)}: Inverse-Wishart + matrix-normal plus a fixed \code{df}.

\item \code{defaultPrior(NormalMvSpec)}: Data-scaled Normal-Inverse-Wishart prior (multivariate).

Control overrides: \code{cLoc} (mean-dispersion multiplier; prior covariance
of \code{mu} is \code{Sigma / kappa0} with \code{kappa0 = 1 / cLoc^2},
default \code{cLoc = 2}) and \code{df0} (inverse-Wishart degrees of freedom,
default \code{d + 2}, must exceed \code{d + 1} for a finite, non-singular
prior covariance on empty components).

\item \code{defaultPrior(NormalGammaMvSpec)}: Normal-Inverse-Wishart prior plus a fixed \code{df}
(default 5, must exceed 2 for a finite component covariance).

\item \code{defaultPrior(NormalUvSpec)}: Data-scaled Normal-Inverse-Gamma prior.

Control overrides: \code{cLoc} (location spread multiplier; prior sd of
\code{mu} ~ \code{cLoc * sd(data)}, default 2), \code{nu0} (InvGamma shape,
default 3, must exceed 2 for a finite prior variance), and \code{concPrior}.

\item \code{defaultPrior(NormalGammaUvSpec)}: Normal-Inverse-Gamma prior plus a fixed \code{df}
(degrees of freedom, default 4, must exceed 2 for a finite component
variance).

\item \code{defaultPrior(PoissonSpec)}: Data-scaled Gamma prior on the Poisson rate
(\code{E[lambda] ~= mean(y)}).

\item \code{defaultPrior(BinomialSpec)}: Data-scaled Beta prior on the success probability.
Requires the number of trials in \code{control$size}.

\item \code{defaultPrior(SkewNormalMvSpec)}: Data-scaled skew-mv-Normal prior.

\item \code{defaultPrior(SkewNormalMvOSpec)}: Adds the Householder angle bound to the mv prior.

\item \code{defaultPrior(SkewIStudentMvSpec)}: Data-scaled skew-mv-IStudent prior.

\item \code{defaultPrior(SkewIStudentMvOSpec)}: Adds the Householder angle bound.

\item \code{defaultPrior(SkewNormalMvOGenSpec)}: Adds the Householder angle box.

\item \code{defaultPrior(SkewIStudentMvOGenSpec)}: Adds the Householder angle box.

\item \code{defaultPrior(StudentTMvSpec)}: Normal-Inverse-Wishart prior plus a fixed \code{df}
(default 5, must exceed 2 for a finite component covariance).

\item \code{defaultPrior(StudentTUvSpec)}: Data-scaled Normal location / Gamma precision prior
  for the Student-t component.

Control overrides: \code{cLoc} (location spread multiplier, default 2),
\code{df} (degrees of freedom, a fixed hyperparameter, default 4, must
exceed 2 for a finite component variance).

}}
