% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonorm-utils.R
\name{.log_omega_gmsnburr}
\alias{.log_omega_gmsnburr}
\title{Compute Log-Omega for GMSNBurr}
\usage{
.log_omega_gmsnburr(alpha, theta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{theta}{Positive shape parameter.}
}
\value{
Numeric value of log(omega).
}
\description{
Internal function for the log normalizing constant of the GMSNBurr
distribution.
}
\keyword{internal}
