% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-EngineConfig.R
\docType{class}
\name{EngineConfig-class}
\alias{EngineConfig-class}
\title{Virtual base class for mixture sampling engines}
\description{
Virtual base class for mixture sampling engines
}
\section{Slots}{

\describe{
\item{\code{name}}{Character scalar identifying the engine.}
}}

