% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-fssn.R,
%   R/dist-fossep.R, R/dist-fsst.R, R/dist-glm-reg.R, R/dist-gmsnburr.R,
%   R/dist-normal-reg.R, R/dist-jfst.R, R/dist-sep.R, R/dist-lep.R,
%   R/dist-msnburr.R, R/dist-mv-reg.R, R/dist-normal-mv.R, R/dist-normal-uv.R,
%   R/dist-poisson-binomial.R, R/dist-skewnormal-mv.R,
%   R/dist-skewnormal-mv-o.R, R/dist-skewistudent-mv.R,
%   R/dist-skewistudent-mv-o.R, R/dist-skewnormal-mv-og.R,
%   R/dist-skewistudent-mv-og.R, R/dist-student-t.R
\name{extractParamTraces}
\alias{extractParamTraces}
\alias{extractParamTraces,FSSNUvSpec-method}
\alias{extractParamTraces,FOSSEPUvSpec-method}
\alias{extractParamTraces,FSSTUvSpec-method}
\alias{extractParamTraces,PoissonRegSpec-method}
\alias{extractParamTraces,BinomialRegSpec-method}
\alias{extractParamTraces,GMSNBurrUvSpec-method}
\alias{extractParamTraces,NormalRegSpec-method}
\alias{extractParamTraces,JFSTUvSpec-method}
\alias{extractParamTraces,SEPUvSpec-method}
\alias{extractParamTraces,LEPUvSpec-method}
\alias{extractParamTraces,MSNBurrUvSpec-method}
\alias{extractParamTraces,MSNBurr2aUvSpec-method}
\alias{extractParamTraces,NormalMvRegSpec-method}
\alias{extractParamTraces,NormalMvSpec-method}
\alias{extractParamTraces,NormalUvSpec-method}
\alias{extractParamTraces,PoissonSpec-method}
\alias{extractParamTraces,BinomialSpec-method}
\alias{extractParamTraces,SkewNormalMvSpec-method}
\alias{extractParamTraces,SkewNormalMvOSpec-method}
\alias{extractParamTraces,SkewIStudentMvSpec-method}
\alias{extractParamTraces,SkewIStudentMvOSpec-method}
\alias{extractParamTraces,SkewNormalMvOGenSpec-method}
\alias{extractParamTraces,SkewIStudentMvOGenSpec-method}
\alias{extractParamTraces,StudentTUvSpec-method}
\title{Parse raw cluster-parameter traces from the MCMC sample matrix}
\usage{
extractParamTraces(spec, samples, L, ...)

\S4method{extractParamTraces}{FSSNUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{FOSSEPUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{FSSTUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{PoissonRegSpec}(spec, samples, L, d = NULL, prior = NULL, ...)

\S4method{extractParamTraces}{BinomialRegSpec}(spec, samples, L, d = NULL, prior = NULL, ...)

\S4method{extractParamTraces}{GMSNBurrUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{NormalRegSpec}(spec, samples, L, d = NULL, prior = NULL, ...)

\S4method{extractParamTraces}{JFSTUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{SEPUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{LEPUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{MSNBurrUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{MSNBurr2aUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{NormalMvRegSpec}(spec, samples, L, d = NULL, prior = NULL, ...)

\S4method{extractParamTraces}{NormalMvSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{NormalUvSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{PoissonSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{BinomialSpec}(spec, samples, L, ...)

\S4method{extractParamTraces}{SkewNormalMvSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{SkewNormalMvOSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{SkewIStudentMvSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{SkewIStudentMvOSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{SkewNormalMvOGenSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{SkewIStudentMvOGenSpec}(spec, samples, L, d = NULL, ...)

\S4method{extractParamTraces}{StudentTUvSpec}(spec, samples, L, ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{samples}{The MCMC sample matrix (iterations x monitored nodes).}

\item{L}{Integer truncation length.}

\item{...}{Reserved for methods.}
}
\value{
A named list of raw parameter traces (matrices/arrays) keyed by the
  logical parameter names of the distribution.
}
\description{
Parse raw cluster-parameter traces from the MCMC sample matrix
}
\section{Functions}{
\itemize{
\item \code{extractParamTraces(FSSNUvSpec)}: Parse mu / sigma / alpha traces.

\item \code{extractParamTraces(FOSSEPUvSpec)}: Parse mu / sigma / alpha / theta traces.

\item \code{extractParamTraces(FSSTUvSpec)}: Parse mu / sigma / alpha / nu traces.

\item \code{extractParamTraces(PoissonRegSpec)}: Parse coefficient traces.

\item \code{extractParamTraces(BinomialRegSpec)}: Parse coefficient traces.

\item \code{extractParamTraces(GMSNBurrUvSpec)}: Parse mu / sigma / alpha / theta traces.

\item \code{extractParamTraces(NormalRegSpec)}: Parse betaTilde (L x p) and s2Tilde (L)
traces; \code{prior$coefNames} (if present) labels the coefficients.

\item \code{extractParamTraces(JFSTUvSpec)}: Parse mu / sigma / alpha / theta traces.

\item \code{extractParamTraces(SEPUvSpec)}: Parse mu / sigma / nu traces.

\item \code{extractParamTraces(LEPUvSpec)}: Parse mu / sigma / nu traces.

\item \code{extractParamTraces(MSNBurrUvSpec)}: Parse mu / sigma / alpha traces.

\item \code{extractParamTraces(MSNBurr2aUvSpec)}: Parse mu / sigma / alpha traces.

\item \code{extractParamTraces(NormalMvRegSpec)}: Parse coefficient and covariance traces.

\item \code{extractParamTraces(NormalMvSpec)}: Parse muTilde (L x d) and covTilde
(L x d x d) traces into arrays.

\item \code{extractParamTraces(NormalUvSpec)}: Parse muTilde / s2Tilde traces.

\item \code{extractParamTraces(PoissonSpec)}: Parse lambda traces.

\item \code{extractParamTraces(BinomialSpec)}: Parse prob traces.

\item \code{extractParamTraces(SkewNormalMvSpec)}: Parse mu / Sigma / gamma traces.

\item \code{extractParamTraces(SkewNormalMvOSpec)}: Parse mu / Sigma / gamma / theta traces.

\item \code{extractParamTraces(SkewIStudentMvSpec)}: Parse mu / Sigma / gamma / nu traces.

\item \code{extractParamTraces(SkewIStudentMvOSpec)}: Parse mu / Sigma / gamma / nu / theta traces.

\item \code{extractParamTraces(SkewNormalMvOGenSpec)}: Parse mu / Sigma / gamma / angle traces.

\item \code{extractParamTraces(SkewIStudentMvOGenSpec)}: Parse mu / Sigma / gamma / nu / angle traces.

\item \code{extractParamTraces(StudentTUvSpec)}: Parse muTilde / tauTilde traces.

}}
\keyword{internal}
