% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{FOSSEPRegSpec-class}
\alias{FOSSEPRegSpec-class}
\alias{FOSSEPRegSpec}
\title{FOSSEP regression specification}
\usage{
FOSSEPRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
FOSSEP regression specification
}
