% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{FSSNRegSpec-class}
\alias{FSSNRegSpec-class}
\alias{FSSNRegSpec}
\title{FSSN regression specification}
\usage{
FSSNRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
FSSN regression specification
}
