% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-fssn.R
\docType{class}
\name{FSSNUvSpec-class}
\alias{FSSNUvSpec-class}
\alias{FSSNUvSpec}
\title{FSSN mixture components (Fernandez-Steel skew Normal)}
\usage{
FSSNUvSpec()
}
\description{
Fernandez-Steel skew-Normal component with location \code{mu}, scale
\code{sigma}, and skewness \code{alpha} (\code{alpha = 1} Normal). A
log-normal prior on \code{alpha} treats left/right skew symmetrically.
Non-conjugate.
}
\references{
Fernandez, C. & Steel, M. F. J. (1998). On Bayesian modeling of fat tails
and skewness. JASA 93, 359--371.
Choir, A. S. (2020). The New Neo-Normal Distributions and their Properties.
}
\keyword{internal}
