% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{FSSTRegSpec-class}
\alias{FSSTRegSpec-class}
\alias{FSSTRegSpec}
\title{FSST regression specification}
\usage{
FSSTRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
FSST regression specification
}
