% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SpatialWeightSpec.R
\name{getAdjacency}
\alias{getAdjacency}
\alias{getAdjacency,SpatialWeightSpec-method}
\title{Adjacency matrix of a spatial weight structure}
\usage{
getAdjacency(spec)

\S4method{getAdjacency}{SpatialWeightSpec}(spec)
}
\arguments{
\item{spec}{A \code{\linkS4class{SpatialWeightSpec}}.}
}
\value{
The (named) numeric adjacency matrix.
}
\description{
Adjacency matrix of a spatial weight structure
}
