% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerDistribution.R
\name{getDistribution}
\alias{getDistribution}
\title{Retrieve a registered distribution by name}
\usage{
getDistribution(name)
}
\arguments{
\item{name}{Character scalar.}
}
\value{
A \code{\linkS4class{DistributionSpec}}.
}
\description{
Retrieve a registered distribution by name
}
