% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{GMSNBurrRegSpec-class}
\alias{GMSNBurrRegSpec-class}
\alias{GMSNBurrRegSpec}
\title{GMSNBurr regression specification}
\usage{
GMSNBurrRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
GMSNBurr regression specification
}
