% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-glm-reg.R,
%   R/dist-normal-reg.R, R/dist-mv-reg.R, R/dist-neonormal-reg.R
\name{isRegressionSpec}
\alias{isRegressionSpec}
\alias{isRegressionSpec,DistributionSpec-method}
\alias{isRegressionSpec,PoissonRegSpec-method}
\alias{isRegressionSpec,BinomialRegSpec-method}
\alias{isRegressionSpec,NormalRegSpec-method}
\alias{isRegressionSpec,NormalMvRegSpec-method}
\alias{isRegressionSpec,MSNBurrRegSpec-method}
\alias{isRegressionSpec,SEPRegSpec-method}
\alias{isRegressionSpec,MSNBurr2aRegSpec-method}
\alias{isRegressionSpec,FSSNRegSpec-method}
\alias{isRegressionSpec,GMSNBurrRegSpec-method}
\alias{isRegressionSpec,LEPRegSpec-method}
\alias{isRegressionSpec,FSSTRegSpec-method}
\alias{isRegressionSpec,FOSSEPRegSpec-method}
\alias{isRegressionSpec,JFSTRegSpec-method}
\title{Is this a regression component spec?}
\usage{
isRegressionSpec(spec, ...)

\S4method{isRegressionSpec}{DistributionSpec}(spec, ...)

\S4method{isRegressionSpec}{PoissonRegSpec}(spec, ...)

\S4method{isRegressionSpec}{BinomialRegSpec}(spec, ...)

\S4method{isRegressionSpec}{NormalRegSpec}(spec, ...)

\S4method{isRegressionSpec}{NormalMvRegSpec}(spec, ...)

\S4method{isRegressionSpec}{MSNBurrRegSpec}(spec)

\S4method{isRegressionSpec}{SEPRegSpec}(spec)

\S4method{isRegressionSpec}{MSNBurr2aRegSpec}(spec)

\S4method{isRegressionSpec}{FSSNRegSpec}(spec)

\S4method{isRegressionSpec}{GMSNBurrRegSpec}(spec)

\S4method{isRegressionSpec}{LEPRegSpec}(spec)

\S4method{isRegressionSpec}{FSSTRegSpec}(spec)

\S4method{isRegressionSpec}{FOSSEPRegSpec}(spec)

\S4method{isRegressionSpec}{JFSTRegSpec}(spec)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{...}{Unused.}
}
\value{
Logical scalar.
}
\description{
Predicate the predict path uses to route to the regression branch. Default
\code{FALSE}; regression specs override it to \code{TRUE}.
}
\section{Functions}{
\itemize{
\item \code{isRegressionSpec(DistributionSpec)}: Default: not a regression spec.

\item \code{isRegressionSpec(PoissonRegSpec)}: Poisson regression is a regression spec.

\item \code{isRegressionSpec(BinomialRegSpec)}: Binomial regression is a regression spec.

\item \code{isRegressionSpec(NormalRegSpec)}: Normal-linear regression is a regression spec.

\item \code{isRegressionSpec(NormalMvRegSpec)}: Multivariate-response regression spec.

\item \code{isRegressionSpec(MSNBurrRegSpec)}: MSNBurr regression is a regression spec.

\item \code{isRegressionSpec(SEPRegSpec)}: SEP regression is a regression spec.

\item \code{isRegressionSpec(MSNBurr2aRegSpec)}: MSNBurr-IIa regression is a regression spec.

\item \code{isRegressionSpec(FSSNRegSpec)}: FSSN regression is a regression spec.

\item \code{isRegressionSpec(GMSNBurrRegSpec)}: GMSNBurr regression is a regression spec.

\item \code{isRegressionSpec(LEPRegSpec)}: LEP regression is a regression spec.

\item \code{isRegressionSpec(FSSTRegSpec)}: FSST regression is a regression spec.

\item \code{isRegressionSpec(FOSSEPRegSpec)}: FOSSEP regression is a regression spec.

\item \code{isRegressionSpec(JFSTRegSpec)}: JFST regression is a regression spec.

}}
