% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{JFSTRegSpec-class}
\alias{JFSTRegSpec-class}
\alias{JFSTRegSpec}
\title{JFST regression specification}
\usage{
JFSTRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
JFST regression specification
}
