% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{LEPRegSpec-class}
\alias{LEPRegSpec-class}
\alias{LEPRegSpec}
\title{LEP regression specification}
\usage{
LEPRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
LEP regression specification
}
