% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-lep.R
\docType{class}
\name{LEPUvSpec-class}
\alias{LEPUvSpec-class}
\alias{LEPUvSpec}
\title{LEP mixture components (exponential power, alternative parameterisation)}
\usage{
LEPUvSpec()
}
\description{
Symmetric exponential-power component with location \code{mu}, scale
\code{sigma}, and shape \code{nu}; an alternative parameterisation to
\code{\link{SEPUvSpec}}. Non-conjugate.
}
\references{
Choir, A. S. (2020). The New Neo-Normal Distributions and their Properties.
Dissertation. Institut Teknologi Sepuluh Nopember.
}
\keyword{internal}
