% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerDistribution.R
\name{listDistributions}
\alias{listDistributions}
\title{List registered distribution names}
\usage{
listDistributions()
}
\value{
Character vector of registered names.
}
\description{
List registered distribution names
}
