% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonorm-utils.R
\name{log1mexp}
\alias{log1mexp}
\title{Numerically Stable log(1 - exp(x))}
\usage{
log1mexp(x)
}
\arguments{
\item{x}{Numeric vector, must be < 0.}
}
\value{
Numeric vector of the same length as `x`.
}
\description{
Computes log(1 - exp(x)) for x < 0. Used in upper-tail probability
computations.
}
\keyword{internal}
