% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MixtureModel.R
\docType{class}
\name{MixtureModel-class}
\alias{MixtureModel-class}
\title{Virtual base class for mixture models}
\description{
Virtual base class for mixture models
}
\section{Slots}{

\describe{
\item{\code{data}}{Numeric data (vector for univariate clustering in v0.1.0).}

\item{\code{distSpec}}{A \code{\linkS4class{DistributionSpec}}.}

\item{\code{engine}}{An \code{\linkS4class{EngineConfig}}.}

\item{\code{Kmax}}{Integer truncation level for the number of components.}

\item{\code{prior}}{A named list of prior hyperparameters.}
}}

