% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{MSNBurr2aRegSpec-class}
\alias{MSNBurr2aRegSpec-class}
\alias{MSNBurr2aRegSpec}
\title{MSNBurr-IIa regression specification}
\usage{
MSNBurr2aRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
MSNBurr-IIa regression specification
}
