% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-msnburr.R
\docType{class}
\name{MSNBurr2aUvSpec-class}
\alias{MSNBurr2aUvSpec-class}
\alias{MSNBurr2aUvSpec}
\title{MSNBurr-IIa mixture components (neo-normal, right-skew capable)}
\usage{
MSNBurr2aUvSpec()
}
\description{
Mirror image of \code{\link{MSNBurrUvSpec}}: the MSNBurr-IIa family
(Iriawan 2000; Choir 2020) accommodates right skew, with \code{alpha = 1}
again the logistic distribution.
}
\references{
Choir, A. S. (2020). The New Neo-Normal Distributions and their Properties.
Dissertation. Institut Teknologi Sepuluh Nopember.
}
\keyword{internal}
