% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{MSNBurrRegSpec-class}
\alias{MSNBurrRegSpec-class}
\alias{MSNBurrRegSpec}
\title{MSNBurr regression specification}
\usage{
MSNBurrRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
A mixture of MSNBurr-IIa regressions: each component is a skewed regression
with its own coefficient vector, scale, and shape.
}
