% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-msnburr.R
\docType{class}
\name{MSNBurrUvSpec-class}
\alias{MSNBurrUvSpec-class}
\alias{MSNBurrUvSpec}
\title{MSNBurr mixture components (neo-normal, left-skew capable)}
\usage{
MSNBurrUvSpec()
}
\description{
Univariate neo-normal component family with location \code{mu}, scale
\code{sigma} and shape \code{alpha} controlling skewness (Iriawan 2000;
Choir 2020). \code{alpha = 1} is the logistic distribution. Non-conjugate;
NIMBLE assigns adaptive samplers to the component parameters.
}
\references{
Iriawan, N. (2000). Computationally Intensive Approaches to Inference in
Neo-Normal Linear Models. PhD Thesis, Curtin University of Technology.

Choir, A. S. (2020). The New Neo-Normal Distributions and their
Properties. Dissertation. Institut Teknologi Sepuluh Nopember.
}
\keyword{internal}
