% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SpatialWeightSpec.R
\name{neighborsOf}
\alias{neighborsOf}
\alias{neighborsOf,SpatialWeightSpec-method}
\title{Neighbours of one region}
\usage{
neighborsOf(spec, region)

\S4method{neighborsOf}{SpatialWeightSpec}(spec, region)
}
\arguments{
\item{spec}{A \code{\linkS4class{SpatialWeightSpec}}.}

\item{region}{A region id (character) or index (integer).}
}
\value{
Character vector of neighbouring region ids.
}
\description{
Neighbours of one region
}
