% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimix-package.R
\docType{package}
\name{nimix-package}
\alias{nimix}
\alias{nimix-package}
\title{nimix: Bayesian Mixture Clustering and Regression with NIMBLE}
\description{
Bayesian mixture modelling built on the \pkg{nimble} platform. The
release implements univariate and multivariate Gaussian mixture clustering
through a Dirichlet Process Mixture (DPM) engine based on the Chinese
Restaurant Process. The package is organised around an extensible S4
\code{\linkS4class{DistributionSpec}} contract so that new component
distributions and (later) a reversible-jump engine can be added without
rewriting existing code.
}
\section{Roadmap (this is v0.2.0)}{

\itemize{
  \item v0.1.0: S4 foundation, \code{\linkS4class{NormalUvSpec}},
    univariate \code{\link{nimixClust}} on the DPM engine.
  \item v0.2.0 (this release): multivariate clustering
    (\code{\linkS4class{NormalMvSpec}}).
  \item v0.3.0: mixture-of-regressions (\code{\link{nimixReg}}).
  \item v0.5.0+: reversible jump MCMC engine.
}
}

\references{
de Valpine, P., Turek, D., Paciorek, C.J., Anderson-Bergman, C.,
Temple Lang, D., & Bodik, R. (2017). Programming with models: writing
statistical algorithms for general model structures with NIMBLE.
\emph{Journal of Computational and Graphical Statistics}, 26(2), 403--413.
\doi{10.1080/10618600.2016.1172487}

Neal, R.M. (2000). Markov chain sampling methods for Dirichlet process
mixture models. \emph{Journal of Computational and Graphical Statistics},
9(2), 249--265. \doi{10.1080/10618600.2000.10474879}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/madsyair/nimix}
  \item \url{https://madsyair.github.io/nimix/}
  \item Report bugs at \url{https://github.com/madsyair/nimix/issues}
}

}
\author{
\strong{Maintainer}: nimix Developers \email{nimix@example.org}

Authors:
\itemize{
  \item nimix Developers \email{nimix@example.org}
}

}
\keyword{internal}
