% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\docType{class}
\name{NormalGammaMvRegSpec-class}
\alias{NormalGammaMvRegSpec-class}
\title{Multivariate-response Normal-Gamma mixture regression (scale mixture)}
\description{
Multivariate-response Normal-Gamma mixture regression (scale mixture)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-gamma-mv-reg"}.}
}}

