% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\name{NormalGammaMvRegSpec}
\alias{NormalGammaMvRegSpec}
\title{Construct a multivariate-response Normal-Gamma regression spec}
\usage{
NormalGammaMvRegSpec()
}
\value{
A \code{\linkS4class{NormalGammaMvRegSpec}}.
}
\description{
Construct a multivariate-response Normal-Gamma regression spec
}
\examples{
spec <- NormalGammaMvRegSpec()
}
