% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-gamma-mv.R
\docType{class}
\name{NormalGammaMvSpec-class}
\alias{NormalGammaMvSpec-class}
\title{Multivariate Normal-Gamma (scale-mixture multivariate-t) component}
\description{
Conjugate scale-mixture representation of a multivariate Student-t component:
identical marginal to \code{\linkS4class{StudentTMvSpec}}, with conjugate
Normal-Inverse-Wishart cluster updates. \code{df} is a fixed hyperparameter.
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-gamma-mv"}.}

\item{\code{paramNames}}{\code{c("mu", "Sigma")}.}
}}

\references{
Andrews, D.F., & Mallows, C.L. (1974). Scale mixtures of normal
distributions. \emph{JRSS-B}, 36(1), 99--102.
\doi{10.1111/j.2517-6161.1974.tb00989.x}

Backlund, E., & Hobert, J.P. (2020). [Gibbs sampling for multivariate linear
regression with errors that are scale mixtures of normals under a conjugate
Normal-Inverse-Wishart prior.]
}
\seealso{
\code{\link{StudentTMvSpec}} for the direct (non-conjugate) path.
}
