% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-gamma-mv.R
\name{NormalGammaMvSpec}
\alias{NormalGammaMvSpec}
\title{Construct a multivariate Normal-Gamma component spec}
\usage{
NormalGammaMvSpec()
}
\value{
A \code{\linkS4class{NormalGammaMvSpec}}.
}
\description{
Construct a multivariate Normal-Gamma component spec
}
\examples{
spec <- NormalGammaMvSpec()
}
