% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-heavytail-reg.R
\docType{class}
\name{NormalGammaRegSpec-class}
\alias{NormalGammaRegSpec-class}
\title{Normal-Gamma mixture regression component (conjugate scale-mixture)}
\description{
Normal-Gamma mixture regression component (conjugate scale-mixture)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-gamma-reg"}.}
}}

