% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-heavytail-reg.R
\name{NormalGammaRegSpec}
\alias{NormalGammaRegSpec}
\title{Construct a Normal-Gamma regression component spec}
\usage{
NormalGammaRegSpec()
}
\value{
A \code{\linkS4class{NormalGammaRegSpec}}.
}
\description{
Construct a Normal-Gamma regression component spec
}
\examples{
spec <- NormalGammaRegSpec()
}
