% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-gamma.R
\name{NormalGammaUvSpec}
\alias{NormalGammaUvSpec}
\title{Construct a univariate Normal-Gamma component spec}
\usage{
NormalGammaUvSpec()
}
\value{
A \code{\linkS4class{NormalGammaUvSpec}}.
}
\description{
Construct a univariate Normal-Gamma component spec
}
\examples{
spec <- NormalGammaUvSpec()
}
