% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\docType{class}
\name{NormalMvRegSpec-class}
\alias{NormalMvRegSpec-class}
\title{Multivariate-response Normal mixture regression}
\description{
Multivariate-response Normal mixture regression
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-mv-reg"}.}
}}

