% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\name{NormalMvRegSpec}
\alias{NormalMvRegSpec}
\title{Construct a multivariate-response Normal regression spec}
\usage{
NormalMvRegSpec()
}
\value{
A \code{\linkS4class{NormalMvRegSpec}}.
}
\description{
Construct a multivariate-response Normal regression spec
}
\examples{
spec <- NormalMvRegSpec()
}
