% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-mv.R
\docType{class}
\name{NormalMvSpec-class}
\alias{NormalMvSpec-class}
\title{Multivariate Gaussian component specification}
\description{
Multivariate Gaussian component specification
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-mv"}.}

\item{\code{paramNames}}{\code{c("mu", "Sigma")}.}

\item{\code{dataDim}}{\code{NA_integer_}: the actual dimension \eqn{d} is taken from
the data at fit time and carried in the prior list.}
}}

\references{
Dellaportas, P., & Papageorgiou, I. (2006). Multivariate mixtures of normals
with unknown number of components. \emph{Statistics and Computing}, 16,
57--68. \doi{10.1007/s11222-006-5338-6}
}
\seealso{
\code{\link{nimixClust}}, \code{\linkS4class{NormalUvSpec}}
}
