% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-mv.R
\name{NormalMvSpec}
\alias{NormalMvSpec}
\title{Construct a multivariate Gaussian component spec}
\usage{
NormalMvSpec()
}
\value{
A \code{\linkS4class{NormalMvSpec}}.
}
\description{
Construct a multivariate Gaussian component spec
}
\examples{
spec <- NormalMvSpec()
}
