% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-reg.R
\docType{class}
\name{NormalRegSpec-class}
\alias{NormalRegSpec-class}
\title{Normal-linear regression component specification}
\description{
Normal-linear regression component specification
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-reg"}.}

\item{\code{paramNames}}{\code{c("beta", "s2")}.}

\item{\code{dataDim}}{\code{1L} (univariate response).}
}}

\references{
Hurn, M., Justel, A., & Robert, C.P. (2003). Estimating mixtures of
regressions. \emph{Journal of Computational and Graphical Statistics},
12(1), 55--79. \doi{10.1198/1061860031329}
}
\seealso{
\code{\link{nimixReg}}
}
