% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-reg.R
\name{NormalRegSpec}
\alias{NormalRegSpec}
\title{Construct a Normal-linear regression component spec}
\usage{
NormalRegSpec()
}
\value{
A \code{\linkS4class{NormalRegSpec}}.
}
\description{
Construct a Normal-linear regression component spec
}
\examples{
spec <- NormalRegSpec()
}
