% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-uv.R
\docType{class}
\name{NormalUvSpec-class}
\alias{NormalUvSpec-class}
\title{Univariate Gaussian component specification}
\description{
Univariate Gaussian component specification
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"normal-uv"}.}

\item{\code{paramNames}}{\code{c("mu", "s2")}.}

\item{\code{priorSpec}}{Filled by \code{\link{defaultPrior}}.}
}}

\references{
Frühwirth-Schnatter, S. (2006). \emph{Finite Mixture and Markov Switching
Models}. Springer. \doi{10.1007/978-0-387-35768-3}

Neal, R.M. (2000). Markov chain sampling methods for Dirichlet process
mixture models. \emph{JCGS}, 9(2), 249--265.
\doi{10.1080/10618600.2000.10474879}
}
\seealso{
\code{\link{nimixClust}}
}
