% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-normal-uv.R
\name{NormalUvSpec}
\alias{NormalUvSpec}
\title{Construct a univariate Gaussian component spec}
\usage{
NormalUvSpec()
}
\value{
A \code{\linkS4class{NormalUvSpec}}.
}
\description{
Construct a univariate Gaussian component spec
}
\examples{
spec <- NormalUvSpec()
}
