% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SpatialWeightSpec.R
\name{nRegions}
\alias{nRegions}
\alias{nRegions,SpatialWeightSpec-method}
\title{Number of regions in a spatial weight structure}
\usage{
nRegions(spec)

\S4method{nRegions}{SpatialWeightSpec}(spec)
}
\arguments{
\item{spec}{A \code{\linkS4class{SpatialWeightSpec}}.}
}
\value{
Integer number of regions.
}
\description{
Number of regions in a spatial weight structure
}
