% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonorm-utils.R
\name{omega_gmsnburr}
\alias{omega_gmsnburr}
\title{Compute Omega for GMSNBurr}
\usage{
omega_gmsnburr(alpha, theta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{theta}{Positive shape parameter.}
}
\value{
Numeric value of omega.
}
\description{
Computes the normalizing constant omega for the GMSNBurr distribution.
}
\keyword{internal}
