% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-glm-reg.R
\docType{class}
\name{PoissonRegSpec-class}
\alias{PoissonRegSpec-class}
\title{Poisson GLM regression component (log link)}
\description{
Poisson GLM regression component (log link)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"poisson-reg"}.}

\item{\code{paramNames}}{\code{c("beta")}.}
}}

