% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-glm-reg.R
\name{PoissonRegSpec}
\alias{PoissonRegSpec}
\title{Construct a Poisson regression component spec}
\usage{
PoissonRegSpec()
}
\value{
A \code{\linkS4class{PoissonRegSpec}}.
}
\description{
Construct a Poisson regression component spec
}
\examples{
spec <- PoissonRegSpec()
}
