% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-poisson-binomial.R
\docType{class}
\name{PoissonSpec-class}
\alias{PoissonSpec-class}
\title{Poisson count component specification}
\description{
Poisson count component specification
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"poisson"}.}

\item{\code{paramNames}}{\code{c("lambda")}.}
}}

