% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-poisson-binomial.R
\name{PoissonSpec}
\alias{PoissonSpec}
\title{Construct a Poisson component spec}
\usage{
PoissonSpec()
}
\value{
A \code{\linkS4class{PoissonSpec}}.
}
\description{
Construct a Poisson component spec
}
\examples{
spec <- PoissonSpec()
}
