% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MixtureModel.R
\docType{class}
\name{RegressionMixModel-class}
\alias{RegressionMixModel-class}
\title{Mixture-of-regressions model (for nimixReg)}
\description{
Mixture-of-regressions model (for nimixReg)
}
\section{Slots}{

\describe{
\item{\code{data}}{Numeric response vector.}

\item{\code{formula}}{The model formula.}

\item{\code{X}}{The numeric design matrix (one row per observation, including the
intercept column when present).}
}}

