% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MixtureModel.R
\name{RegressionMixModel}
\alias{RegressionMixModel}
\title{Construct a RegressionMixModel}
\usage{
RegressionMixModel(data, X, formula, distSpec, engine, Kmax, prior)
}
\arguments{
\item{data}{Numeric response vector.}

\item{X}{Numeric design matrix.}

\item{formula}{The model formula.}

\item{distSpec}{A \code{\linkS4class{DistributionSpec}}.}

\item{engine}{An \code{\linkS4class{EngineConfig}}.}

\item{Kmax}{Integer truncation level.}

\item{prior}{Named list of prior hyperparameters.}
}
\value{
A \code{RegressionMixModel}.
}
\description{
Construct a RegressionMixModel
}
\keyword{internal}
