% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labelswitch.R
\name{relabel}
\alias{relabel}
\alias{relabel,FitResult-method}
\title{Correct label switching in a fitted mixture}
\usage{
relabel(fit, method = "ECR-ITERATIVE-1", ...)

\S4method{relabel}{FitResult}(fit, method = "ECR-ITERATIVE-1", ...)
}
\arguments{
\item{fit}{A \code{\linkS4class{FitResult}}.}

\item{method}{One of \code{"ECR-ITERATIVE-1"} (default) or \code{"ECR"}.
The default is the pivot-free iterative ECR; \code{"ECR"} uses the
highest-posterior allocation as the pivot.}

\item{...}{Reserved.}
}
\value{
The \code{fit} with its \code{relabeled} slot populated.
}
\description{
Post-hoc relabelling of MCMC output so that per-component posterior
summaries are meaningful. Conditions on the modal number of occupied
clusters, then applies an algorithm from the \pkg{label.switching} package.
}
\section{Functions}{
\itemize{
\item \code{relabel(FitResult)}: Relabelling for a fitted result.

}}
\references{
Papastamoulis, P., & Iliopoulos, G. (2010). An artificial allocations based
solution to the label switching problem. \emph{JCGS}, 19(2), 313--331.
\doi{10.1198/jcgs.2010.09008}

Papastamoulis, P. (2016). label.switching: An R package for dealing with the
label switching problem in MCMC outputs. \emph{JSS, Code Snippets}, 69(1).
\doi{10.18637/jss.v069.c01}
}
