% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-neonormal-reg.R
\docType{class}
\name{SEPRegSpec-class}
\alias{SEPRegSpec-class}
\alias{SEPRegSpec}
\title{SEP regression specification}
\usage{
SEPRegSpec()
}
\value{
An object used internally by \code{\link{nimixReg}}.
}
\description{
SEP regression specification
}
