% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewistudent-mv-og.R
\docType{class}
\name{SkewIStudentMvOGenSpec-class}
\alias{SkewIStudentMvOGenSpec-class}
\alias{SkewIStudentMvOGenSpec}
\title{Skew mv independent-Student components with estimated O, general dimension}
\usage{
SkewIStudentMvOGenSpec()
}
\description{
As \code{\link{SkewIStudentMvOSpec-class}} but for any \eqn{m \ge 3}, using
the general Householder parameterisation and the canonicalisation of
\code{\link{canonicaliseO}}. Reached via
\code{distribution = "skewistudent-mv-o"} for data with more than two
columns. Experimental; see \code{\link{SkewNormalMvOGenSpec-class}} for how
to read \code{gamma} and \code{O} after canonicalisation.
}
\references{
Ferreira & Steel (2007), Statistica Sinica 17, 505--529.
}
\keyword{internal}
