% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewistudent-mv-o.R
\docType{class}
\name{SkewIStudentMvOSpec-class}
\alias{SkewIStudentMvOSpec-class}
\alias{SkewIStudentMvOSpec}
\title{Skew mv independent-Student components with estimated orthogonal factor}
\usage{
SkewIStudentMvOSpec()
}
\description{
As \code{\link{SkewIStudentMvSpec-class}}, but the orthogonal factor of
\eqn{A = OU} is estimated through the Householder angle \code{theta} with a
uniform prior on \eqn{(-\pi/8, \pi/8)} (FS restriction (8)). Bivariate data
only. \code{theta} is identified only through the skewness.
}
\references{
Ferreira & Steel (2007), Statistica Sinica 17, 505--529.
}
\keyword{internal}
