% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewnormal-mv-og.R
\docType{class}
\name{SkewNormalMvOGenSpec-class}
\alias{SkewNormalMvOGenSpec-class}
\alias{SkewNormalMvOGenSpec}
\title{Skew multivariate Normal components with estimated O, general dimension}
\usage{
SkewNormalMvOGenSpec()
}
\description{
As \code{\link{SkewNormalMvOSpec-class}} but for any \eqn{m \ge 3}: the
orthogonal factor of \eqn{A = OU} is parameterised by \eqn{m(m-1)/2}
Householder angles, sampled on the FS angle box, and FS's identifiability
restriction (8) is applied as a post-hoc canonicalisation of the posterior
draws (see \code{\link{canonicaliseO}}). Reached via
\code{distribution = "skewnormal-mv-o"} when the data have more than two
columns.
}
\references{
Ferreira & Steel (2007), Statistica Sinica 17, 505--529.
}
\keyword{internal}
