% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-skewnormal-mv-o.R
\docType{class}
\name{SkewNormalMvOSpec-class}
\alias{SkewNormalMvOSpec-class}
\alias{SkewNormalMvOSpec}
\title{Skew multivariate Normal components with estimated orthogonal factor}
\usage{
SkewNormalMvOSpec()
}
\description{
As \code{\link{SkewNormalMvSpec-class}}, but the orthogonal factor of
\eqn{A = OU} is estimated through the Householder angle \code{theta} with a
uniform prior on \eqn{(-\pi/8, \pi/8)} (FS restriction (8)). Bivariate data
only. \code{theta} is identified only through the skewness: at
\code{gamma = 1} the density is invariant in \code{theta}.
}
\references{
Ferreira & Steel (2007), Statistica Sinica 17, 505--529.
}
\keyword{internal}
