% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\docType{class}
\name{StudentTMvRegSpec-class}
\alias{StudentTMvRegSpec-class}
\title{Multivariate-response Student-t mixture regression (direct density)}
\description{
Multivariate-response Student-t mixture regression (direct density)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"student-t-mv-reg"}.}
}}

