% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-mv-reg.R
\name{StudentTMvRegSpec}
\alias{StudentTMvRegSpec}
\title{Construct a multivariate-response Student-t regression spec}
\usage{
StudentTMvRegSpec()
}
\value{
A \code{\linkS4class{StudentTMvRegSpec}}.
}
\description{
Construct a multivariate-response Student-t regression spec
}
\examples{
spec <- StudentTMvRegSpec()
}
