% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-student-t-mv.R
\docType{class}
\name{StudentTMvSpec-class}
\alias{StudentTMvSpec-class}
\title{Multivariate Student-t component specification (direct density)}
\description{
Direct multivariate-t kernel via a user-defined NIMBLE distribution. Same
marginal as \code{\linkS4class{NormalGammaMvSpec}} but non-conjugate. \code{df}
is a fixed hyperparameter.
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"student-t-mv"}.}

\item{\code{paramNames}}{\code{c("mu", "Sigma")}.}
}}

\references{
Lange, K.L., Little, R.J.A., & Taylor, J.M.G. (1989). Robust statistical
modeling using the t distribution. \emph{JASA}, 84(408), 881--896.
\doi{10.1080/01621459.1989.10478852}
}
\seealso{
\code{\link{NormalGammaMvSpec}} for the conjugate path.
}
