% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-student-t-mv.R
\name{StudentTMvSpec}
\alias{StudentTMvSpec}
\title{Construct a multivariate Student-t component spec}
\usage{
StudentTMvSpec()
}
\value{
A \code{\linkS4class{StudentTMvSpec}}.
}
\description{
Construct a multivariate Student-t component spec
}
\examples{
spec <- StudentTMvSpec()
}
