% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-heavytail-reg.R
\docType{class}
\name{StudentTRegSpec-class}
\alias{StudentTRegSpec-class}
\title{Student-t mixture regression component (direct t density)}
\description{
Student-t mixture regression component (direct t density)
}
\section{Slots}{

\describe{
\item{\code{name}}{Fixed to \code{"student-t-reg"}.}
}}

