% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-heavytail-reg.R
\name{StudentTRegSpec}
\alias{StudentTRegSpec}
\title{Construct a Student-t regression component spec}
\usage{
StudentTRegSpec()
}
\value{
A \code{\linkS4class{StudentTRegSpec}}.
}
\description{
Construct a Student-t regression component spec
}
\examples{
spec <- StudentTRegSpec()
}
