% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-student-t.R
\name{StudentTUvSpec}
\alias{StudentTUvSpec}
\title{Construct a univariate Student-t component spec}
\usage{
StudentTUvSpec()
}
\value{
A \code{\linkS4class{StudentTUvSpec}}.
}
\description{
Construct a univariate Student-t component spec
}
\examples{
spec <- StudentTUvSpec()
}
