% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DistributionSpec.R, R/dist-normal-reg.R,
%   R/dist-glm-reg.R, R/dist-mv-reg.R
\name{isRegressionSpec}
\alias{isRegressionSpec}
\alias{isRegressionSpec,DistributionSpec-method}
\alias{isRegressionSpec,NormalRegSpec-method}
\alias{isRegressionSpec,PoissonRegSpec-method}
\alias{isRegressionSpec,BinomialRegSpec-method}
\alias{isRegressionSpec,NormalMvRegSpec-method}
\title{Is this a regression component spec?}
\usage{
isRegressionSpec(spec, ...)

\S4method{isRegressionSpec}{DistributionSpec}(spec, ...)

\S4method{isRegressionSpec}{NormalRegSpec}(spec, ...)

\S4method{isRegressionSpec}{PoissonRegSpec}(spec, ...)

\S4method{isRegressionSpec}{BinomialRegSpec}(spec, ...)

\S4method{isRegressionSpec}{NormalMvRegSpec}(spec, ...)
}
\arguments{
\item{spec}{A \code{\linkS4class{DistributionSpec}}.}

\item{...}{Unused.}
}
\value{
Logical scalar.
}
\description{
Predicate the predict path uses to route to the regression branch. Default
\code{FALSE}; regression specs override it to \code{TRUE}.
}
\section{Functions}{
\itemize{
\item \code{isRegressionSpec(DistributionSpec)}: Default: not a regression spec.

\item \code{isRegressionSpec(NormalRegSpec)}: Normal-linear regression is a regression spec.

\item \code{isRegressionSpec(PoissonRegSpec)}: Poisson regression is a regression spec.

\item \code{isRegressionSpec(BinomialRegSpec)}: Binomial regression is a regression spec.

\item \code{isRegressionSpec(NormalMvRegSpec)}: Multivariate-response regression spec.

}}
